fix: Moved websockets code to notifications
